Add GC pause and CPU usage
