--- lib/memory-usage.js
@@ -1,14 +1,0 @@
-const metrics = require('./metrics');
-
-const LOOP = 5000;
-
-function instrument() {
-  setInterval(() => {
-    var usage = process.memoryUsage();
-    metrics.observeMemory(usage.rss, usage.heapTotal, usage.heapUsed); 
-  }, LOOP);
-}
-
-module.exports = {
-  instrument: instrument
-}
