+++ lib/metrics.js
@@ -14,6 +14,11 @@ const metric = {
         total: new client.Gauge('nodejs_memory_heap_total_bytes', 'the V8 heap total in bytes'),
         used: new client.Gauge('nodejs_memory_heap_used_bytes', 'the V8 heap used in bytes')
       }
+    },
+    cpu: new client.Counter('nodejs_cpu_usage_microseconds_total', 'microseconds spent on CPU',['mode']),
+    gc: {
+      time: new client.Counter('nodejs_gc_pause_nanoseconds_total', 'nanoseconds spent in gc pause',['gctype']),
+      reclaimed: new client.Counter('nodejs_gc_reclaimed_bytes_total', 'nanoseconds spent in gc pause',['gctype'])
     }
   }
 }
@@ -44,9 +49,32 @@ function observeLag(start, interval) {
   metric.node.lag.set(lag);
 };
 
+function observeCPU(data) {
+  metric.node.cpu.labels("user").inc(data.user);
+  metric.node.cpu.labels("system").inc(data.system);
+};
+
+const gcTypes = {
+  0: 'unknown',
+  1: 'minor',
+  2: 'major',
+  3: 'both',
+}
+
+var totalReclaimed = 0
+function observeGC(stats) {
+  metric.node.gc.time.labels(gcTypes[stats.gctype]).inc(stats.pause);
+  if (stats.diff.usedHeapSize < 0) {
+    totalReclaimed += stats.diff.usedHeapSize * -1;
+    metric.node.gc.reclaimed.labels(gcTypes[stats.gctype]).inc(totalReclaimed);
+  }
+};
+
 module.exports = {
   observe: observe,
   observeLag: observeLag,
+  observeGC: observeGC,
+  observeCPU: observeCPU,
   observeMemory: observeMemory,
   summary: client.register.metrics
 };
